Backend/requirements.txt
alembic==1.10.3
asgiref==3.7.2
attrs==22.2.0
beautifulsoup4==4.9.3
black==23.3.0
build==0.10.0
CacheControl==0.12.11
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==3.1.0
cleo==2.0.1
click==8.1.3
colorama==0.4.6
crashtest==0.4.1
distlib==0.3.6
Django==4.2.4
django-cors-headers==4.2.0
django-environ==0.9.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
dulwich==0.21.3
filelock==3.10.0
flake8==6.0.0
flake8-black==0.3.6
geckodriver-autoinstaller==0.1.0
greenlet==2.0.2
html5lib==1.1
idna==2.8
importlib-metadata==6.1.0
installer==0.7.0
internetdownloadmanager==0.0.2
isort==5.12.0
jaraco.classes==3.2.3
jsonschema==4.17.3
keyring==23.13.1
lockfile==0.12.2
Mako==1.2.4
MarkupSafe==2.1.2
mccabe==0.7.0
more-itertools==9.1.0
msgpack==1.0.5
mypy-extensions==1.0.0
mysql-connector-python==8.0.32
numpy==1.25.1
packaging==23.0
pandas==2.0.3
pathspec==0.11.1
pexpect==4.8.0
pkginfo==1.9.6
platformdirs==2.6.2
poetry==1.4.1
poetry-core==1.5.2
poetry-plugin-export==1.3.0
psycopg2==2.9.7
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==2.8.0
pyproject_hooks==1.0.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-decouple==3.7
python-dotenv==1.0.0
pytz==2023.3
pywin32-ctypes==0.2.0
rapidfuzz==2.13.7
requests==2.28.2
requests-toolbelt==0.10.1
selenium==3.141.0
shellingham==1.5.0.post1
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.9
sqlparse==0.4.4
termcolor==1.1.0
tomlkit==0.11.6
tqdm==4.48.2
trove-classifiers==2023.3.9
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
virtualenv==20.21.0
webencodings==0.5.1
zipp==3.15.0
